LPS-93654 Migrate portal-remote-jaxrs-whiteboard-test to use arquillian bridge.
