+++ modules/apps/portal-remote/portal-remote-jaxrs-whiteboard-test/bnd.bnd
@@ -1,5 +1,3 @@
-Bundle-Activator: com.liferay.portal.remote.jaxrs.whiteboard.test.internal.activator.RemoteJaxrsWhiteboardTestBundleActivator
 Bundle-Name: Liferay Portal Remote JAXRS Whiteboard Test
 Bundle-SymbolicName: com.liferay.portal.remote.jaxrs.whiteboard.test
 Bundle-Version: 1.0.0
-Private-Package: com.liferay.portal.remote.jaxrs.whiteboard.test.internal.*
