+++ modules/apps/portal-remote/portal-remote-jaxrs-whiteboard-test/build.gradle
@@ -1,11 +1,6 @@
 dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1"
-	compileOnly group: "org.apache.httpcomponents", name: "httpclient", version: "4.3.5"
-	compileOnly group: "org.jboss.arquillian.junit", name: "arquillian-junit-container", version: "1.1.10.Final"
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
-	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
-
-	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-container-liferay", version: "1.0.7"
-	testIntegrationCompile group: "com.liferay.arquillian", name: "com.liferay.arquillian.arquillian-deployment-generator-bnd", version: "1.0.4"
+	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
+	testIntegrationCompile group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1"
+	testIntegrationCompile group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
 }
