--- modules/apps/portal-remote/portal-remote-jaxrs-whiteboard-test/src/main/java/com/liferay/portal/remote/jaxrs/whiteboard/test/internal/service/Greeter.java
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.remote.jaxrs.whiteboard.test.internal.service;
-
-import java.util.Collections;
-import java.util.Set;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Application;
-
-/**
- * @author Carlos Sierra Andrés
- */
-public class Greeter extends Application {
-
-	@Override
-	public Set<Object> getSingletons() {
-		return Collections.<Object>singleton(this);
-	}
-
-	@GET
-	@Path("/sayHello")
-	@Produces("text/plain")
-	public String sayHello() {
-		return "Hello.";
-	}
-
-}
